feat(cli,io-api): Applied project namespace to modules
