--- cli/src/main/java/me/concision/extractor/output/writers/single/ListFormatWriter.java
@@ -1,37 +1,0 @@
-package me.concision.extractor.output.writers.single;
-
-import me.concision.extractor.Extractor;
-import me.concision.extractor.api.PackageJsonifier;
-import me.concision.extractor.api.PackageParser;
-import me.concision.extractor.output.FormatType;
-import org.bson.Document;
-import org.json.JSONArray;
-
-/**
- * See {@link FormatType#LIST}
- *
- * @author Concision
- */
-public class ListFormatWriter extends SingleRecordFormatWriter {
-    private final JSONArray list = new JSONArray();
-
-    @Override
-    public void close() {
-        outputStream.print(list.toString(2));
-        super.close();
-    }
-
-    @Override
-    public void publish(Extractor extractor, PackageParser.PackageRecord record) {
-        Document document = new Document();
-
-        document.put("path", record.fullPath());
-        if (extractor.args().rawMode) {
-            document.put("package", record.contents());
-        } else {
-            document.put("package", PackageJsonifier.parse(record.contents()));
-        }
-
-        list.put(document);
-    }
-}
