--- io-api/src/main/java/me/concision/extractor/api/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Decaching API for raw manipulation of packages.
- * Various algorithms are adapted from https://github.com/cheahjs/WFPackageParser
- *
- * @author Concision
- */
-package me.concision.extractor.api;
