+++ io-api/src/main/java/me/concision/unnamed/packages/ioapi/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Decaching API for raw manipulation of packages.
+ * Various algorithms are adapted from https://github.com/cheahjs/WFPackageParser
+ *
+ * @author Concision
+ */
+package me.concision.unnamed.packages.ioapi;
